% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.dendrogram.twinspan.R
\name{as.dendrogram.twinspan}
\alias{as.dendrogram.twinspan}
\title{Extract Species or Quadrat Dendrograms}
\usage{
\method{as.dendrogram}{twinspan}(
  object,
  height = c("level", "chi", "eigen"),
  what = c("quadrat", "species"),
  subset,
  ...
)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}

\item{height}{Use either division levels (\code{"level"}), total
Chi-squares (\code{"chi"}) or eigenvalues of first axis
(\code{"eigen"}) of division as dendrogram heights.}

\item{what}{Return either a \code{"quadrat"} or \code{"species"}
dendrogram.}

\item{subset}{A logical vector or indices that select a subset of
items to a dendrogram.}

\item{\dots}{Other parameters to functions.}
}
\value{
A \code{\link[stats]{dendrogram}} object.
}
\description{
Function extracts the species or quadrat classification as a
hierarchic \code{\link[stats]{dendrogram}}. Unlike \code{hclust},
twinspan dendrograms show all quadrats or species, and the final
divisions are polytomous.
}
\details{
The dendrogram heights are levels of divisions, or total
Chi-squares of divisions and groups, or eigenvalues of divisions
depending on argument \code{height}. Terminal groups have no
eigenvalues, because they were not considered for division. For
them the method uses arbitrary value that for a group of \eqn{n}
units is proportion \eqn{(n-1)/n} of the height of mother
division. Chi-squares are evaluated also for terminal groups.
There is no guarantee that eigenvalues or Chi-squares decrease in
divisions, and there may be reversals where lower levels are higher
than their mother groups, and the plotted trees can be messy and
unreadable. Chi-squares decrease more monotonically than
eigenvalues of first axis.

Function allows selecting a subset of items.

\R{} has a wealth of functions to handle and display
dendrograms. See \code{\link[stats]{dendrogram}} for general
description. There is even stronger support in packages (for
instance, \CRANpkg{dendextend}).

The terminal groups of \code{\link{twinspan}} trees are not binary,
but may have several elements (quadrats, species).  In
\code{\link[stats]{dendrogram}} plots, it is best to set
\code{type="triangle"} for nicer looking trees.
}
\examples{

## Large datasets are difficult to show in dendrograms, but
#'subset' allows slicing dendrogram.

data(ahti)
tw <- twinspan(ahti)
## cut at level 2 into groups
cl2 <- cut(tw, level=2)
table(cl2)
den <- as.dendrogram(tw, subset = cl2 == 4)
str(den, max.level = 4)
plot(den, type = "triangle")
den <- as.dendrogram(tw, height="chi", subset = cl2 == 4)
plot(den, type = "triangle")
## show only most frequent species
freq <- colSums(ahti > 0)
den <- as.dendrogram(tw, what = "species", height="chi", subset = freq > 16)
plot(den, type = "triangle")

}
\seealso{
\code{\link{as.hclust.twinspan}}, \code{\link{dendrogram}}.
}
