% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.hclust.twinspan.R
\name{as.hclust.twinspan}
\alias{as.hclust.twinspan}
\title{Extract Quadrat Grouping as Hierarchical Cluster Tree}
\usage{
\method{as.hclust}{twinspan}(x, ...)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{\dots}{Other parameters to the function (ignored).}
}
\description{
Function extracts quadrat classification as an \code{\link{hclust}}
object. The terminal items are the quadrat classes, but quadrats
are not shown: \code{\link{hclust}} cannot handle polytomies that
are needed to display individual quadrats.  Use
\code{\link{as.dendrogram}} to show the single items or species
classification.
}
