% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodspec.R
\name{goodspec}
\alias{goodspec}
\title{Most Diagnostic and Characteristic Species}
\usage{
goodspec(object, what = c("both", "indicator", "leading"), label = FALSE)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}

\item{what}{Select either species that are used as
\code{"indicator"} pseudospecies or most abundant and common
(or \code{"leading"}) species of each group species
classification or \code{"both"}.}

\item{label}{Return species labels instead of indices.}
}
\value{
Either indices of species or their name labels.
}
\description{
Function can return (1) species that are used as indicators in
\code{twinspan} classification, or (2) most abundant and common
species (known as leading species) of each final \code{twispan}
class, or both. The default is to return indices that can be used
in subsetting data, but it is also possible to return the species
labels.
}
\examples{
## similar to plot(tw, what="species") but labels classes by their
## leading species
data(ahti)
tw <- twinspan(ahti)
plot(as.dendrogram(tw, what = "species", subset = goodspec(tw, "leading")))

}
\seealso{
Functions \code{\link{image.twinspan}},
    \code{\link{twinheat}} and \code{\link{twintable}} can use the
    function to select species.
}
