% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.twinspan.R
\name{image.twinspan}
\alias{image.twinspan}
\title{Image of Quadrat and Species Classification and Average Abundances in Cells}
\usage{
\method{image}{twinspan}(
  x,
  leadingspecies = FALSE,
  reorder = FALSE,
  height = c("level", "chi"),
  ...
)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{leadingspecies}{Show averages of leading species (most
abundant, ties broken by frequency) of each cluster instead of
averages of all species of the group.}

\item{reorder}{Do not use the original \code{twinspan} ordering of
quadrats and species, but reorder the dendrograms. Setting this
\code{TRUE} will use the first correspondence analysis axis to
reorder the tree which often improves the diagonal
structure. With this option, it is also possible to use
arguments \code{Colv} and \code{Rowv} vectors to reorder the
dendrogram by their values. The length of such a vector must
correspond to the number of row or column classes.}

\item{height}{Use either division levels (\code{"level"}) or total
Chi-squares of division (\code{"chi"}) as heights of internal
nodes in the trees boarding the image (see
\code{\link{as.hclust.twinspan}}).}

\item{\dots}{Other arguments passed to \code{\link[vegan]{tabasco}}
and further to \code{\link[stats]{heatmap}}.}
}
\value{
Function returns invisibly the matrix of mean abundances
    used to produce the graph.
}
\description{
Function draws an \code{\link{image}} of mean pseudospecies values
in cells of final quadrat and species classifications together with
corresponding cluster trees.
}
\details{
The function is based on \pkg{vegan} function
\code{\link[vegan]{tabasco}}.  Mean abundances in cells are shown
by colours which can be modified by the user (see
\code{\link[vegan]{tabasco}}).  The mean abundances are either the
average values of all species of the species group (default), or
the average of the most abundant species of the species group (tied
abundances broken by species frequency) if
\code{leadingspecies=TRUE}. The species groups are rows and quadrat
groups are columns. These are labelled by their decimal group
numbers with information of group sizes, or species by their name
if only leading species are shown. These numbers and the order of
the groups are similar as in hierarchic cluster trees produced by
\code{\link{as.hclust.twinspan}} and
\code{\link{summary.twinspan}}.  Function also adds quadrat
classification tree on the top and species classification tree on
the left margin (see \code{\link{as.hclust.twinspan}}). These trees
show either the levels of hierarchy or the heterogeneities of
divided groups depending on the argument \code{height}.

Function \code{\link{twinheat}} is a visually similar alternative
that displays individual quadrats and species instead of
\code{twinspan} classes.
}
\examples{

data(ahti)
tw <- twinspan(ahti)
image(tw)
im <- image(tw, height = "chi", leading = TRUE)
## image returns invisibly data
head(im)

}
\seealso{
\code{\link[vegan]{tabasco}} and \code{\link{heatmap}} for
    basic functionality. The default colour scheme is based on
    \code{\link{heat.colors}}, but better shemes can be constructed
    (\CRANpkg{viridis} package provides clear schemes). The
    dendrograms are based on \code{\link{as.hclust.twinspan}}.
    Functions \code{\link{twintable}} and \code{\link{twinheat}}
    provide alternatives that can show all quadrats and all
    species.
}
