% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indscores.R
\name{indscores}
\alias{indscores}
\alias{summary.indscores}
\title{Indicator Scores of Pseudospecies in a Division}
\usage{
indscores(object, division)

\method{summary}{indscores}(object, indlimit = 0.2, ...)
}
\arguments{
\item{object}{\code{twinspan} result object.}

\item{division}{Number code of \code{twinspan} division for quadrats.}

\item{indlimit}{Minimum absolute value of returned indicator scores}

\item{\dots}{Other arguments passed to the function (ignored).}
}
\description{
Function returns the indicator scores that \code{\link{twinspan}}
uses to select indicator pseudospecies.
}
\details{
Indicator scores are differences of relative frequencies of
pseudospecies in the posivite (right) and negative (left)
branches. Perfect indicator pseudopecies has value 1 or -1 and
occurs on every quadrat of one branch and in none of the other
branch. Indicator pseudospecies of division are selected by the
indicator scores, but not all obviously best species occur as
indicators. One basic species can be listed only once using its
best pseudospecies, and redundant pseudospecies are not listed.
}
\examples{
data(ahti)
tw <- twinspan(ahti)
summary(indscores(tw, 1))
summary(indscores(tw, 2))
summary(indscores(tw, 3))

}
\seealso{
Function \code{\link{summary.twinspan}} lists the used
    indicator pseudospecies in each division.
}
