% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.hclust.twinspan.R
\name{plot.twinspan}
\alias{plot.twinspan}
\title{Plot Classificaton Tree of Quadrats}
\usage{
\method{plot}{twinspan}(x, main = "Twinspan Dendrogram", ...)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{main}{Main title of the plot.}

\item{\dots}{Other parameters passed to \code{\link{plot}} and
\code{\link[vegan]{ordilabel}}.}
}
\description{
Plot Classificaton Tree of Quadrats
}
