% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinord.R
\name{twinord}
\alias{twinord}
\alias{plot.twinord}
\title{Ordination Demonstrating Twinspan Division}
\usage{
twinord(object, division)

\method{plot}{twinord}(
  x,
  quadrats = c("text", "points", "score", "none"),
  pseudospecies = c("none", "all", "indicators", "valuable"),
  value = 0.25,
  qcol = c("blue", "red"),
  ...
)
}
\arguments{
\item{object}{\code{twinspan} result object.}

\item{division}{Number code of \code{twinspan} division for quadrats.}

\item{x}{\code{twinord} result object.}

\item{quadrats}{Plot quadrats as text, points or with their
indicator score or none.}

\item{pseudospecies}{Plot text for none, all, indicator
pseudospecies or valuable species exceeding \code{value}.}

\item{value}{Limit of absolute indicator value for displayed
pseudospecies.}

\item{qcol}{Colour of quadrat text in negative and positive groups.}

\item{\dots}{Other arguments passed to graphical functions.}
}
\value{
Function returns an object of class \code{"twinord"} which
    inherits from a \pkg{vegan} \code{\link[vegan]{decorana}}
    object, but adds the following auxiliary elements:

\describe{
\item{class}{Two classes into which this division was split.}
\item{indscores}{Indicator scores for quadrats from \code{\link{indscores}}.}
\item{indicators}{Name labels of indicator pseudospecies used to split this
  class from \code{\link{twinspan}} (see \code{\link{summary.twinspan}}).}
\item{indvals}{Indicator values for pseudospecies from
  \code{\link{indvalues}}.}
\item{division}{\code{twinspan} division.}
}
}
\description{
Function performs orthogonal correspondence analysis with
downweighting for quadrats in one division (\pkg{vegan} function
\code{\link[vegan]{decorana}}). This analysis is similar as the one
performed in \code{twinspan} when it splits the division. The
results can be used to visualize and analyse the division. Function
returns the ordination object with auxiliary information on
subdivision, quadrat indicator scores, indicator pseudospecies and
indicator values of pseudospecies.
}
