% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsform.R
\name{twinsform}
\alias{twinsform}
\title{Transform Data for Correspondence Analysis like twinspan}
\usage{
twinsform(x, cutlevels = c(0, 2, 5, 10, 20), subset)
}
\arguments{
\item{x}{Input (community) data.}

\item{cutlevels}{Cut levels used to split quantitative data into
binary pseudospecies.}

\item{subset}{Logical vector or indices that select a subset of
quadrats (sampling units).}
}
\description{
Function transforms data so that Correspondence Analysis gives the
same result as in \code{\link{twinspan}}.
}
