% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinspan.R
\name{twinspan}
\alias{twinspan}
\title{Two-Way Indicator Species Analysis}
\usage{
twinspan(x, cutlevels = c(0, 2, 5, 10, 20))
}
\arguments{
\item{x}{Input data, usually a species community data set where
columns give the species and rows the sampling units.}

\item{cutlevels}{Cut levels used to split quantitative data into
binary pseudospecies.}
}
\description{
Two-Way Indicator Analysis (TWINSPAN) is a divisive classification
method that works by splitting first Correspondence-Analysis into
two, and then recursively working with each of this split
subsets. The function is actually much more complicated: see
Details.
}
