% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twintotalchi.R
\name{totalchi}
\alias{totalchi}
\alias{twintotalchi}
\title{Total Scaled Chi-square of All Divisions and Groups}
\usage{
totalchi(x)

twintotalchi(x, what = c("quadrat", "species"))
}
\arguments{
\item{x}{A matrix of non-negative data for \code{totalchi} or a
\code{\link{twinspan}} result object for \code{twintotalchi}.}

\item{what}{Analyse \code{quadrat} or \code{species}
classification.}
}
\description{
Total scaled Chi-square is the same as the sum of all eigenvalues
in Correspondence Analysis. It is a measure of heterogeneity of a
set of data. The function finds this measure for all divisions and
terminal groups in \code{\link{twinspan}}
}
\details{
Function first reconstructs the data as it was internally used in
\code{\link{twinspan}} using functions \code{\link{twin2stack}} and
\code{\link{twin2specstack}}. The scaled Chi-square is calculated
with support function \code{totalchi} that can be called
independently for any matrix with non-negative data. The scaling in
Chi-square means that data are standardized to unit sum before
calculating the actual Chi-square. This is often called the sum of
all eigenvalues in Correspondence Analysis, but no eigenvalues are
evaluated in these functions, only their potential sum.
}
\examples{

data(ahti)
tw <- twinspan(ahti)
twintotalchi(tw)
}
\seealso{
The basic functions are \code{\link{twin2stack}} and
    \code{\link{twin2specstack}} that construct the data matrices.
}
