% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.dendrogram.twinspan.R
\name{as.dendrogram.twinspan}
\alias{as.dendrogram.twinspan}
\title{Extract Species or Quadrat Dendrograms}
\usage{
\method{as.dendrogram}{twinspan}(object, what = c("quadrat", "species"),
  eigenheight = FALSE, ...)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}

\item{what}{Return either a \code{"quadrat"} or \code{"species"}
dendrogram.}

\item{eigenheight}{Use eigenvalues of division as dendrogram
heights. The default is to use division level as heights. NB.,
there is no guarantee that eigenvalues decrease in divisions,
and there may be reversals where lower levels are higher than
lower levels.}
}
\value{
A \code{\link[stats]{dendrogram}} object.
}
\description{
Function extracts the species or quadrat classification as a
hierarchic \code{\link[stats]{dendrogram}}.
}
\details{
\R{} has a wealth of functions to handle and display
dendrograms. See \code{\link[stats]{dendrogram}} for general
description. There is even stronger support in packages (for
instance, \CRANpkg{dendextend}).

The \code{twinspan} dendrogram is not a completely binary tree, but
there are several final units (quadrats, species). In
\code{\link[stats]{dendrogram}} plots, it is best to set
\code{type="triangle"} for nicer looking plots. This is even more
important if the dendrogram heights are based on division
eigenvalues, because then the groups heights are regularly
reversed.
}
