% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.dendrogram.twinspan.R
\name{as.dendrogram.twinspan}
\alias{as.dendrogram.twinspan}
\title{Extract Species or Quadrat Dendrograms}
\usage{
\method{as.dendrogram}{twinspan}(object, what = c("quadrat", "species"),
  eigenheight = FALSE, ...)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}

\item{what}{Return either a \code{"quadrat"} or \code{"species"}
dendrogram.}

\item{eigenheight}{Use eigenvalues of division as dendrogram
heights. Terminal groups have no eigenvalues, because they were
not considered for division. For them use arbitrary value that
for a group of \eqn{n} units is proportion \eqn{n/(n-1)} of the
height of mother division.  There is no guarantee that
eigenvalues decrease in divisions, and there may be reversals
where lower levels are higher than their mother groups, and the
plotted trees can be messy and unreadable.}

\item{\dots}{Other parameters to functions.}
}
\value{
A \code{\link[stats]{dendrogram}} object.
}
\description{
Function extracts the species or quadrat classification as a
hierarchic \code{\link[stats]{dendrogram}}.
}
\details{
\R{} has a wealth of functions to handle and display
dendrograms. See \code{\link[stats]{dendrogram}} for general
description. There is even stronger support in packages (for
instance, \CRANpkg{dendextend}).

The terminal groups of \code{\link{twinspan}} trees are not binary,
but may have several elements (quadrats, species).  In
\code{\link[stats]{dendrogram}} plots, it is best to set
\code{type="triangle"} for nicer looking trees.
}
