% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.hclust.twinspan.R
\encoding{UTF-8}
\name{as.hclust.twinspan}
\alias{as.hclust.twinspan}
\title{Extract twinspan Grouping as Hierarchical Cluster Tree}
\usage{
\method{as.hclust}{twinspan}(x, what = c("quadrat", "species"), height = c("level", "chi"), ...)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{what}{Extract \code{"quadrat"} or \code{"species"}
classification tree.}

\item{height}{Use either division levels (\code{"level"}) or total
Chi-squares of division (\code{"chi"}) as heights of internal
nodes in the tree.}

\item{\dots}{Other parameters to the function (ignored).}
}
\value{
an \code{\link{hclust}} object amended with labels for
    internal nodes (\code{nodelabels}).
}
\description{
Function extracts classification as an \code{\link{hclust}}
object. The terminal items are the final groups, but quadrats or
species are not shown: \code{\link{hclust}} cannot handle
polytomies that are needed to display group members.  Use
\code{\link{as.dendrogram}} to show the single items. The group ID
number and number of items in the terminal group are used as group
names and are displayed in plots.
}
\details{
Function can return either a tree showing the \code{twinspan}
hierarchy or showing the heterogeneity of each group or
division. In the first case, all divisions and groups at a certain
level of hierarchy are at the same height, but in the latter the
divisions are at the height defined by their heterogeneity. The
criterion of heterogeneity is the total chi-square (also known as
inertia) of the matrix that \code{\link{twinspan}} internally uses
in that division (see \code{\link{twintotalchi}}). This tree gives
the visual presentation of the modified method of Roleček et
al. (2009).

When tree heights are based on heterogeneity, subgroups can be more
heterogeneous than their parent group. These appear as reversed
branches in the tree. A warning is issued for each such case.
}
\examples{

data(ahti)
tw <- twinspan(ahti)
plot(as.hclust(tw, "species"))
cl <- as.hclust(tw)
## plot and 8 groups by hierarchy level
plot(cl)
rect.hclust(cl, 8)
## plot and 8 groups by heterogeneity
cl <- as.hclust(tw, height="chi")
plot(cl)
rect.hclust(cl, 8)



}
\references{
Roleček, J, Tichý, L., Zelený, D. & Chytrý, M. (2009). Modified
TWINSPAN classification in which the hierarchy respects cluster
heterogeneity. \emph{J Veg Sci} 20: 596--602.
}
\seealso{
\code{\link{as.dendrogram.twinspan}} provides an
    alternative which also shows the sampling units (quadrats or
    species). The result is based \code{\link{hclust}} and can be
    handled with its support
    functions. \code{\link{plot.twinspan}},
    \code{\link{image.twinspan}} display the tree. Function
    \code{\link{cut.twinspan}} cuts the tree by a level of
    hierarchy, and \code{\link{cuth}} by heterogeneity for original
    sampling units (quadrats, species).
}
