% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.twinspan.R
\name{cut.twinspan}
\alias{cut.twinspan}
\title{Return twinspan Classification at Given Level}
\usage{
\method{cut}{twinspan}(x, level, what = c("quadrat", "species"), ...)
}
\arguments{
\item{x}{\code{twinspan} result.}

\item{level}{Level of hierarchy for classification. If missing, the
highest lvel used in the object will be returned.}

\item{what}{Return either a \code{"quadrat"} or \code{"species"}
classification vector.}

\item{\dots}{Other parameters (ignored).}
}
\value{
A vector of class numbers for the given level of hierarchy
    using Twinspan identifiers. A mother class can be found through
    integer division by 2 for a given class identifier.
}
\description{
Function returns a vector of \code{twinspan} classes at given level
of hierarcy for quadrats.
}
