% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvals.twinspan.R
\name{eigenvals.twinspan}
\alias{eigenvals.twinspan}
\alias{eigenvals}
\title{Eigenvalues of twinspan Divisions}
\usage{
\method{eigenvals}{twinspan}(x, what = c("quadrat", "species"), ...)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{what}{Return eigenvalues of \code{"quadrat"} or
\code{"species"} divisions.}

\item{\dots}{Other arguments (ignored).}
}
\description{
Eigenvalues of twinspan Divisions
}
\examples{

data(ahti)
tw <- twinspan(ahti)
eigenvals(tw)

}
