% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.twinspan.R
\name{image.twinspan}
\alias{image.twinspan}
\title{Image of Quadrat and Species Classification and Average Abundances in Cells}
\usage{
\method{image}{twinspan}(x, ...)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{\dots}{Other arguments passed to \code{\link[vegan]{tabasco}}
and further to \code{\link[stats]{heatmap}}.}
}
\value{
Function returns invisibly the matrix of mean abundances
    used to produce the graph.
}
\description{
Function draws an \code{\link{image}} of mean pseudospecies values
in cells of final quadrat and species classifications together with
corresponding cluster trees.
}
\details{
The function is based on \CRANpkg{vegan} function
\code{\link[vegan]{tabasco}}.  Mean abundances in cells are shown
by colours which can be modified by the user (see
\code{\link[vegan]{tabasco}}). The species groups are rows and
quadrat groups are columns. These are labelled by their decimal
group numbers with information of group sizes. These numbers and
the order of the groups are similar as in hierarchic cluster trees
produced by \code{\link{as.hclust.twinspan}} and
\code{\link{summary.twinspan}}. Function also adds quadrat
classification tree on the top and species classification tree on
the left margin (see \code{\link{as.hclust.twinspan}}).
}
