% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indscores.R
\name{indscores}
\alias{indscores}
\alias{plot.indscores}
\title{Indicator Scores for Quadrats}
\usage{
indscores(object, division)

\method{plot}{indscores}(x, col = c("red", "blue"), ...)
}
\arguments{
\item{object}{\code{twinspan} result object.}

\item{division}{Number code of \code{twinspan} division for quadrats.}

\item{x}{\code{indscores} result object.}

\item{col}{Colours for negative (left) and positive (right) groups.}

\item{...}{Other arguments passed to the function (ignored).}
}
\description{
Indicator scores are the sum of negative and positive pseudospecies
indicators on each quadrat.
}
