% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassified.R
\name{misclassified}
\alias{misclassified}
\title{Report Misclassified Quadrats}
\usage{
misclassified(x, level, verbose = TRUE, binname = FALSE)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{level}{Only consider misclassification down to this level.}

\item{verbose}{If \code{FALSE}, returns only the index of
misclassified quadrats.}

\item{binname}{Use binary labels instead of decimal numbers for
classes and divergent division in verbose object.}
}
\value{
If \code{verbose=TRUE}, the function returns an object of class
\code{"misclassified"} with following elements.
\describe{
\item{index}{Index of misclassified quadrats.}
\item{labels}{Labels (names) of quadrats.}
\item{class}{Final classification from \code{\link{twinspan}}.}
\item{predicted}{Final classification from \code{\link{predict.twinspan}}.}
\item{division}{The division where the misclassification occurred.}
}

With \code{verbose=FALSE}, the function returns a logical vector
that is \code{TRUE} for misclassified quadrats.
}
\description{
\code{\link{twinspan}} bases its quadrat classification primarily
on ordination axis. In some cases this is in conflict with the
classification derived from indicator pseudospecies. This function
identifies these cases.
}
\details{
The function compares the final \code{\link{twinspan}}
classification (from \code{\link{cut.twinspan}}) and the
classification predicted from indicator pseudospecies (from
\code{\link{predict.twinspan}}). If these two differ, the quadrat
is \dQuote{misclassified}. With \code{verbose=FALSE}, the function
returns a logical vector that is \code{TRUE} for misclassified
quadrats. In default, it also returns the names, twinspan classes
and predicted classes of misclassified quadrats, and the division
where the misclassification occurred and classifications
diverged. The divisions and their numbers can be seen in
\code{\link{summary.twinspan}} and \code{\link{plot.twinspan}}.
}
\examples{

data(ahti)
tw <- twinspan(ahti)
misclassified(tw)
## see the ID numbers of divisions
plot(tw)
## only look at misclassifications at first two levels
misclassified(tw, level = 2)

}
\seealso{
The basic functions are \code{\link{cut.twinspan}} and
    \code{\link{predict.twinspan}}. You can see the division
    numbers with \code{\link{summary.twinspan}} (with indicator
    pseudospecies) and in \code{\link{plot.twinspan}}.
}
