% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.hclust.twinspan.R
\name{plot.twinspan}
\alias{plot.twinspan}
\title{Plot Classification Tree}
\usage{
\method{plot}{twinspan}(
  x,
  what = c("quadrat", "species"),
  height = c("level", "chi"),
  main = "Twinspan Dendrogram",
  ...
)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{what}{Plot \code{"quadrat"} or \code{"species"}
classification tree.}

\item{height}{Use either division levels (\code{"level"}) or total
Chi-squares of division (\code{"chi"}) as heights of internal
nodes in the tree.}

\item{main}{Main title of the plot.}

\item{\dots}{Other parameters passed to \code{\link{plot}} and
\code{\link[vegan]{ordilabel}}.}
}
\description{
Function displays the classification tree.
}
\details{
The internal nodes are labelled by the numbers of division. These
are the same numbers as used in \code{\link{summary.twinspan}} and
returned by \code{\link{cut.twinspan}} or
\code{\link{predict.twinspan}} for the same classification
level. For terminal groups the plot shows the number of the group
and the number of items (quadrats or species) in the group. For
division number \eqn{k}, its daughter divisions or groups are
\eqn{2k}{2*k} and \eqn{2k+1}{2*k+1}. The tree is similar as a plot
of \code{\link{as.hclust.twinspan}}, but adds numbers of internal
nodes.
}
\examples{

data(ahti)
tw <- twinspan(ahti)
plot(tw, "species")

}
\seealso{
\code{\link{summary.twinspan}} for similar textual
    presentation also showing the items (quadrats, species) in
    terminal groups. \CRANpkg{vegan} function
    \code{\link[vegan]{scores.hclust}} can extract the coordinates
    of internal (or terminal nodes), and
    \code{\link[vegan]{ordilabel}} is used add the labels on
    internal nodes.
}
