% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.twinspan.R
\name{predict.twinspan}
\alias{predict.twinspan}
\title{Predict Class Membership of Quadrats}
\usage{
\method{predict}{twinspan}(object, newdata, level, ...)
}
\arguments{
\item{object}{\code{twinspan} result object.}

\item{newdata}{Data used in prediction. The species will be matched
by their names, and the pseudospecies are based on the
\code{cutlevels} used in the original \code{twinspan} model.}

\item{level}{Level of hierarchy of classification. If missing, the
prediction is made to the highest level of classification.}

\item{\dots}{Other parameters passed to the function (ignored).}
}
\description{
Function predicts the class membership for each quadrat using the
reported indicator pseudospecies and limit for the indicator score
for the \dQuote{positive} (left) group.
}
\details{
The \code{twinspan} classification is based on splicing polarized
ordination axis, and the reported indicator pseudospecies only
indicate the decisions in each division, and do not necessarily
give the same classification: The original classification cannot be
necessarily found when giving the original data as
\code{newdata}. In the original TWINSPAN this called
misclassification.
}
