% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.twinspan.R
\name{summary.twinspan}
\alias{summary.twinspan}
\title{Summary of twinspan Classification}
\usage{
\method{summary}{twinspan}(object, what = c("quadrat", "species"), binname = FALSE, maxitems, ...)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}

\item{what}{Summarize either quadrat or species classification.}

\item{binname}{Use binary labels for divisions instead of decimal numbers.}

\item{maxitems}{Maximum number of items (members) listed for terminal groups.}

\item{\dots}{Other arguments (ignored).}
}
\value{
The function returns nothing. It only prints the result
    object in a human-readable way.
}
\description{
The function gives a compact summary of divisions with indicator
species and items in final classification. The output gives the
same essential information as the printed output of TWINSPAN batch
program, but in more compact form.
}
\details{
For each division, \code{summary} prints the eigenvalue. For
quadrat divisions, it also prints the indicator pseudospecies with
their signs, followed by \code{<} and the lowest indicator score
for the \sQuote{positive} (right) group. If the indicator score is
below this value, follow the summary to the next item at the lower
level, and if the indicator score is at the limit or higher, follow
to the second alternative.  For division number \eqn{k}, the next
items are either \eqn{2k}{2*k} (\sQuote{negative} group) or
\eqn{2k+1}{2*k+1} (\sQuote{positive} group). Function
\code{\link{plot.twinspan}} displays the division numbers in a
classification tree. Function \code{\link{indvalues}} returns the
used numeric indicator scores, and lists also the pseudospecies
that were not among official indicators (repeated pseudospecies of
the same basic species, redundant pseudospecies that do not improve
the classification).

For terminal groups, the function gives the size of the group and
lists its elements (quadrats or species).
}
\examples{
data(ahti)
tw <- twinspan(ahti)
summary(tw, maxitems = 6)
summary(tw, "species")

}
\seealso{
\code{\link{plot.twinspan}} displays the same structure
    visually.  Function \code{\link{predict.twinspan}} follows the
    summary strcture to predict the classification with indicator
    pseudospecies. \code{\link{indvalues}} gives the numeric
    indicator values used to select the indicator pseudospecies.
}
