% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.twinspan.R
\name{summary.twinspan}
\alias{summary.twinspan}
\title{Summary of Twinspan Classification}
\usage{
\method{summary}{twinspan}(object, what = c("quadrat", "species"), ...)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}

\item{what}{Summarize either quadrat or species classification.}

\item{\dots}{Other arguments (ignored).}
}
\value{
The function returns nothing. It only prints the result
    object in a human-readable way.
}
\description{
The function gives a compact summary of divisions with indicator
species and items in final classification. The output gives the
same essential information as the printed output of TWINSPAN batch
program, but in more compact form.
}
\details{
For each division, \code{summary} prints the eigenvalue, and for
quadrat divisions the indicator pseudospecies with their signs,
followed by \code{<} and the lowest score for the \dQuote{positive}
group. If the indicator score is below this value, follow the list
to the next item at the lower level, and to the second alternative
at the same level if the indicator score is at the limit or
higher. For division \eqn{k}, the next items are \eqn{2k}
(\dQuote{negative} group) or \eqn{2k+1} (\dQuote{positive}
group). For terminal groups, the function gives the size of the
group and lists its elements (quadrats or species).
}
\examples{
data(ahti)
tw <- twinspan(ahti)
summary(tw)
summary(tw, "species")

}
