% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsform.R
\name{twin2stack}
\alias{twin2stack}
\alias{twin2mat}
\title{Extract Transformed Input Data from TWINSPAN result}
\usage{
twin2stack(x, subset, downweight = FALSE)

twin2mat(x)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{subset}{Select a subset of quadrats.}

\item{downweight}{Downweight infrequent pseudospecies.}
}
\description{
Function \code{twin2stack} extracts the pseudospecies matrix, where
columns are pseudospecies with their cutlevels. This is similar to
the file generated with \code{\link{twinsform}}. The default is to
return a binary matrix, where data entries are eiter \eqn{0} or
\eqn{1}. However, it is also possible to extract a subset of data
with downweighting allowing scrutiny of \code{\link{twinspan}}
divisions.

Function \code{twin2mat} extracts data file with pseudospecies
transformation. Columns are original species, and entries are
abundances after pseudospecies transformation. This is similar as
the output from \CRANpkg{vegan} function
\code{\link[vegan]{coverscale}} with similar cut levels and
argument \code{character=FALSE}.
}
\examples{

data(ahti)
dim(ahti)
range(ahti)
tw <- twinspan(ahti)
x <- twin2mat(tw)
dim(x)
range(x)
colnames(x)
x <- twin2stack(tw)
dim(x)
range(x)
colnames(x)
## Inspect division 4
x <- twin2stack(tw, subset = cut(tw, 2) == 4, downweight = TRUE)
## need vegan for correspondence analysis
if (suppressPackageStartupMessages(require("vegan"))) {
cca(x)
}

}
\seealso{
\code{\link{twinsform}}, \code{\link[vegan]{coverscale}}.
}
