% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsform.R
\name{twin2stack}
\alias{twin2stack}
\alias{twin2mat}
\title{Extract Transformed Input Data from twinspan result}
\usage{
twin2stack(x, subset, downweight = FALSE)

twin2mat(x)
}
\arguments{
\item{x}{\code{\link{twinspan}} result object.}

\item{subset}{Select a subset of quadrats.}

\item{downweight}{Downweight infrequent pseudospecies.}
}
\description{
Functions extract the data \code{\link{twinspan}} used in its
analysis, and allow reproducing the internal ordination and
inspecting the \code{twinspan} divisions.
}
\details{
Function \code{twin2stack} extracts the pseudospecies matrix, where
columns are pseudospecies with their cutlevels. This is similar to
the file generated with \code{\link{twinsform}}. The default is to
return a binary matrix, where data entries are eiter \eqn{0} or
\eqn{1}. Alternatively, it is possible to extract a subset of data
with downweighting allowing scrutiny of \code{\link{twinspan}}
divisions.

Function \code{twin2mat} extracts data file with pseudospecies
transformation. Columns are original species, and entries are
abundances after pseudospecies transformation. This is similar as
the output from \CRANpkg{vegan} function
\code{\link[vegan]{coverscale}} with similar cut levels and
argument \code{character=FALSE}. These data were not analysed in
\code{\link{twinspan}}, but these are the data tabulated with
\code{\link{twintable}}.
}
\examples{

data(ahti)
dim(ahti)
range(ahti)
tw <- twinspan(ahti)
x <- twin2mat(tw)
dim(x)
range(x)
colnames(x)
x <- twin2stack(tw)
dim(x)
range(x)
colnames(x)
## Inspect division 4
x <- twin2stack(tw, subset = cut(tw, 2) == 4, downweight = TRUE)
## need vegan for correspondence analysis
if (suppressPackageStartupMessages(require("vegan"))) {
cca(x)
}

}
\seealso{
For original data set instead of \code{\link{twinspan}}
    result, functions \code{\link{twinsform}} and
    \code{\link[vegan]{coverscale}} are analogous to
    \code{twin2stack} and \code{twin2mat}.
}
