% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinheat.R
\name{twinheat}
\alias{twinheat}
\title{Display twinspan Classification as a Heatmap}
\usage{
twinheat(
  object,
  maxspp,
  goodspecies,
  subset,
  height = "level",
  Rowv = FALSE,
  Colv = FALSE,
  ...
)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}

\item{maxspp}{Maximum number of most abundant species
displayed. The abundance is estimated with pseudospecies cut
levels. The default is to show all species.}

\item{goodspecies}{Select \dQuote{good species} for
tabulation. These are either species that were used as
indicator pseudospecies (\code{"indicator"}), or most abundant
species in each final species group breaking ties with
frequency (\code{"leading"}), or \code{"both"} (default). The
abundance is estimated after pseudospecies transformation for
all quadrats and cannot be used together with \code{maxspp}.}

\item{subset}{Select a subset of quadrats.}

\item{height}{Use either division levels (\code{"level"}), total
Chi-squares (\code{"chi"}) or eigenvalues of first axis of
division (\code{"eigen"}) as dendrogram heights.}

\item{Rowv, Colv}{Reorder dendrograms. If \code{FALSE} use original
\code{twinspan} order. If \code{TRUE} order by the first axis
of CCA. Alternatively, the arguments can be vectors that are
used to reorder the dendrogram.}

\item{\dots}{Other arguments passed to \code{\link[vegan]{tabasco}}.}
}
\description{
Display \code{twinspan} classification as a \code{link{heatmap}}
with quadrat and species dendrograms using \pkg{vegan} function
\code{\link[vegan]{tabasco}}.
}
\seealso{
\code{\link{twintable}}, \code{\link{image.twinspan}},
    \code{\link[vegan]{tabasco}}.
}
