% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinord.R
\name{twinord}
\alias{twinord}
\title{Ordination Demonstrating Twinspan Division}
\usage{
twinord(object, division)
}
\arguments{
\item{object}{\code{twinspan} result object.}

\item{division}{Number code of \code{twinspan} division for quadrats.}
}
\value{
Function returns an object of class \code{"twinord"} with elements:
\describe{
\item{ordination}{Orthogonal correspondence analysis solution from
  \code{\link[vegan]{decorana}} (\pkg{vegan} package).}
\item{class}{Two classes into which this division was split.}
\item{indscores}{Indicator scores for quadrats from \code{\link{indscores}}.}
\item{indicators}{Name labels of indicator pseudospecies used to split this
  class from \code{\link{twinspan}} (see \code{\link{summary.twinspan}}).}
\item{indvals}{Indicator values for pseudospecies from
  \code{\link{indvalues}}.}
}
}
\description{
Function performs orthogonal correspondence analysis with
downweighting for quadrats in one division (\pkg{vegan} function
\code{\link[vegan]{decorana}}). This analysis is similar as the one
performed in \code{twinspan} when it splits the division. The
results can be used to visualize and analyse the division. Function
returns the ordination object with auxiliary information on
subdivision, quadrat indicator scores, indicator pseudospecies and
indicator values of pseudospecies.
}
