% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsform.R
\name{twinsform}
\alias{twinsform}
\title{Transform Data for Correspondence Analysis like twinspan}
\usage{
twinsform(x, cutlevels = c(0, 2, 5, 10, 20), subset)
}
\arguments{
\item{x}{Input (community) data.}

\item{cutlevels}{Cut levels used to split quantitative data into
binary pseudospecies.}

\item{subset}{Logical vector or indices that select a subset of
quadrats (sampling units).}
}
\value{
A stacked matrix of downweighted pseudospecies.
}
\description{
Function transforms data so that Correspondence Analysis gives the
same result as in \code{\link{twinspan}}.
}
\details{
In \code{\link{twinspan}}, quantitative species data are split into
binary (0/1) pseudospecies by \code{cutlevels}. All these
pseudospecies are stacked as columns in a new data set. Rare
pseudospecies that occur at lower frequency than 0.2 are
\code{\link[vegan]{downweight}}ed to reduce the importance of rare
species or rare abundance levels in correspondence analysis. When
analysed with correspondence analysis (e.g.,
\code{\link[vegan]{cca}}, \code{\link[vegan]{decorana}} with option
\code{ira=1}), this will give the same eigenvalue and ordination as
used in \code{\link{twinspan}}. When a \code{subset} of a
\code{\link{twinspan}} class is used, correspondence analysis of
subdivision of the class can be obtained. The results are not
absolutely similar, mainly because correspondence analysis is more
approximate in \code{\link{twinspan}} than in dedicated
functions. Moreover, \code{\link{twinspan}} adds a small value
(0.01) to all abundance classes, and does not use weights smaller
than 0.01 which can influence results when the number of quadrats
is above 100.
}
