% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsform.R
\name{twinsform}
\alias{twinsform}
\title{Transform Data for Correspondence Analysis like twinspan}
\usage{
twinsform(x, cutlevels = c(0, 2, 5, 10, 20), subset, downweight = TRUE)
}
\arguments{
\item{x}{Input (community) data.}

\item{cutlevels}{Cut levels used to split quantitative data into
binary pseudospecies.}

\item{subset}{Logical vector or indices that select a subset of
quadrats (sampling units).}

\item{downweight}{Downweight result similarly as in
\code{\link[vegan]{decorana}}. Downweighting is needed to
replicate the process in \code{twinspan}, but it can be left
out when we only want to have a stacked data set for other
uses.}
}
\value{
A stacked matrix of downweighted pseudospecies.
}
\description{
Function transforms data so that Correspondence Analysis gives the
same result as in \code{\link{twinspan}}.
}
\details{
In \code{\link{twinspan}}, quantitative species data are split into
binary (0/1) pseudospecies by \code{cutlevels}. All these
pseudospecies are stacked as columns in a new data set. Rare
pseudospecies that occur at lower frequency than 0.2 are
\code{\link[vegan]{downweight}}ed to reduce the importance of rare
species or rare abundance levels in correspondence analysis. When
analysed with correspondence analysis (e.g.,
\code{\link[vegan]{cca}}, \code{\link[vegan]{decorana}} with option
\code{ira=1}), this will give the same eigenvalue and ordination as
used in \code{\link{twinspan}}. When a \code{subset} of a
\code{\link{twinspan}} class is used, correspondence analysis of
subdivision of the class can be obtained.
}
\examples{

data(ahti)
tahti <- twinsform(ahti)
colnames(tahti)
## needs vegan for correspondence analysis
if (suppressPackageStartupMessages(require("vegan"))) {
decorana(tahti, ira=1)
}
## similar first eigenvalue
eigenvals(twinspan(ahti))

}
\seealso{
\code{\link[vegan]{downweight}} in \CRANpkg{vegan}: this
    function is often used with Detrended Correspondence Analysis
    (\code{\link[vegan]{decorana}}). However, the implementation is
    slightly different in TWINSPAN, and weights differ slightly.
}
