% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twintable.R
\name{twintable}
\alias{twintable}
\title{Community Table Ordered by Twinspan Classification}
\usage{
twintable(object)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}
}
\description{
Prints a community table of pseudospecies ordered by
\code{\link{twinspan}} classification.
}
