% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twintable.R
\name{twintable}
\alias{twintable}
\title{Community Table Ordered by Twinspan Classification}
\usage{
twintable(object, maxspp, goodspecies, subset)
}
\arguments{
\item{object}{\code{\link{twinspan}} result object.}

\item{maxspp}{Maximum number of most abundant species
displayed. The abundance is estimated with pseudospecies cut
levels. The default is to show all species.}

\item{goodspecies}{Select \dQuote{good species} for
tabulation. These are either species that were used as
indicator pseudospecies (\code{"indicator"}), or most abundant
species in each final species group breaking ties with
frequency (\code{"leading"}), or \code{"both"} (default). The
abundance is estimated after pseudospecies transformation for
all quadrats and cannot be used together with \code{maxspp}.}

\item{subset}{Select a subset of quadrats.}
}
\value{
Function returns invisibly the data that
    \code{\link[vegan]{vegemite}} used. This data can also be
    exported or used as any other data set.
}
\description{
Prints a community table of pseudospecies ordered by
\code{\link{twinspan}} classification.
}
\details{
Function prints a compact community table of pseudospecies
values. The table is ordered by clustering both species and
quadrats similarly as in \code{\link{summary.twinspan}} or in plot
of \code{\link{as.dendrogram.twinspan}}. The classification of each
quadrat and species is shown by a sequence of \code{0} and \code{1}
indicating division of each level. This string is binary
presentation of the decimal class number without the leading
\code{1}.

Only one character is used for each abundance, and the table is
very compact. However, large tables can be divided over several
pages or screen windows. The width of the displayed table is
controlled by \R{} option \code{width} (see
\code{\link{options}}). It is possible to select only a
\code{subset} of the quadrats for tabulation giving narrower
tables. The number of species can be reduced by setting the maximum
number of most abundant species, or alternatively, by restricting
tabulation only to \dQuote{good species} which are the most
abundant species of each species group (ties broken by species
frequency), or species used as indicators, or both.
}
\examples{

data(ahti)
tw <- twinspan(ahti)
## complete table would be large, but we show subset of group 4
twintable(tw, subset = twingroup(tw, 4), goodspecies = "both")

}
